requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Django==4.2.27
exceptiongroup==1.3.1
google-auth==2.43.0
google-genai==1.47.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
playsound==1.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
requests==2.32.5
rsa==4.9.1
SpeechRecognition==3.14.4
sqlparse==0.5.4
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1